coredumpctl: move acquire_journal() and focus() to coredumpctl-journal.c
