+++ src/coredump/meson.build
@@ -20,6 +20,7 @@ systemd_coredump_export_sources = files(
 coredumpctl_sources = files(
         'coredumpctl.c',
         'coredumpctl-info.c',
+        'coredumpctl-journal.c',
         'coredumpctl-util.c',
 )
